feat: add TelemetryProvider (#2640)
