+++ apps/web/src/components/Common/Layout.tsx
@@ -1,13 +1,10 @@
 import GlobalAlerts from '@components/Shared/GlobalAlerts';
 import BottomNavigation from '@components/Shared/Navbar/BottomNavigation';
-import FingerprintJS from '@fingerprintjs/fingerprintjs';
 import getIsAuthTokensAvailable from '@lib/getIsAuthTokensAvailable';
 import getToastOptions from '@lib/getToastOptions';
 import resetAuthData from '@lib/resetAuthData';
-import { IS_MAINNET, MIXPANEL_ENABLED, MIXPANEL_TOKEN } from 'data/constants';
 import type { Profile } from 'lens';
 import { useUserProfilesQuery } from 'lens';
-import mixpanel from 'mixpanel-browser';
 import Head from 'next/head';
 import { useTheme } from 'next-themes';
 import type { FC, ReactNode } from 'react';
@@ -15,7 +12,6 @@ import { useEffect } from 'react';
 import { Toaster } from 'react-hot-toast';
 import { CHAIN_ID } from 'src/constants';
 import { useAppPersistStore, useAppStore } from 'src/store/app';
-import { useFingerprintStore } from 'src/store/fingerprint';
 import { useAccount, useDisconnect, useNetwork } from 'wagmi';
 
 import GlobalModals from '../Shared/GlobalModals';
@@ -23,14 +19,7 @@ import Loading from '../Shared/Loading';
 import Navbar from '../Shared/Navbar';
 import useIsMounted from '../utils/hooks/useIsMounted';
 import { useDisconnectXmtp } from '../utils/hooks/useXmtpClient';
-
-if (MIXPANEL_ENABLED) {
-  mixpanel.init(MIXPANEL_TOKEN, {
-    ignore_dnt: true,
-    api_host: '/collect',
-    batch_requests: false
-  });
-}
+import TelemetryProvider from './TelemetryProvider';
 
 interface LayoutProps {
   children: ReactNode;
@@ -42,11 +31,8 @@ const Layout: FC<LayoutProps> = ({ children }) => {
   const setUserSigNonce = useAppStore((state) => state.setUserSigNonce);
   const currentProfile = useAppStore((state) => state.currentProfile);
   const setCurrentProfile = useAppStore((state) => state.setCurrentProfile);
-  const setIsPro = useAppStore((state) => state.setIsPro);
   const profileId = useAppPersistStore((state) => state.profileId);
   const setProfileId = useAppPersistStore((state) => state.setProfileId);
-  const fingerprint = useFingerprintStore((state) => state.fingerprint);
-  const setFingerprint = useFingerprintStore((state) => state.setFingerprint);
 
   const { mounted } = useIsMounted();
   const { address } = useAccount();
@@ -99,49 +85,10 @@ const Layout: FC<LayoutProps> = ({ children }) => {
     }
   };
 
-  const saveFingerprint = async () => {
-    const fp = await FingerprintJS.load();
-    const { visitorId } = await fp.get();
-    setFingerprint(visitorId);
-  };
-
   useEffect(() => {
     validateAuthentication();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [address, chain, disconnect, profileId]);
-
-  useEffect(() => {
-    saveFingerprint();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  // set pro status
-  useEffect(() => {
-    if (currentProfile?.id && currentProfile?.id === '0x0d') {
-      if (IS_MAINNET) {
-        setIsPro(true);
-      } else {
-        setIsPro(true);
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentProfile?.id]);
-
-  // Mixpanel identify
-  useEffect(() => {
-    if (MIXPANEL_ENABLED && currentProfile?.id && fingerprint) {
-      mixpanel.identify(currentProfile?.id);
-      mixpanel.people.set({
-        $name: currentProfile?.handle,
-        $fingerprint: fingerprint,
-        $last_active: new Date()
-      });
-      mixpanel.people.set_once({
-        $created_at: new Date()
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentProfile?.id]);
 
   if (loading || !mounted) {
     return <Loading />;
@@ -153,6 +100,7 @@ const Layout: FC<LayoutProps> = ({ children }) => {
         <meta name="theme-color" content={resolvedTheme === 'dark' ? '#1b1b1d' : '#ffffff'} />
       </Head>
       <Toaster position="bottom-right" toastOptions={getToastOptions(resolvedTheme)} />
+      <TelemetryProvider />
       <GlobalModals />
       <GlobalAlerts />
       <div className="flex min-h-screen flex-col pb-14 md:pb-0">
